src/product_fetcher.py
"""
Product Fetcher Module
Fetches trending/high-commission products from TikTok Shop
"""
import requests
import json
import time
import logging
from typing import List, Dict, Optional
from pathlib import Path
from bs4 import BeautifulSoup
import random

logger = logging.getLogger(__name__)


class ProductFetcher:
    """Handles fetching products from TikTok Shop"""

    def __init__(self, credentials: Dict, filters: Dict):
        self.credentials = credentials
        self.filters = filters
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        })

    def fetch_trending_products(self, limit: int = 20) -> List[Dict]:
        """
        Fetch trending products from TikTok Shop

        NOTE: TikTok Shop API requires official partnership.
        This is a placeholder showing the structure.
        You'll need to:
        1. Apply for TikTok Shop API access
        2. Use their official SDK
        3. Or use web scraping (risky, against TOS)
        """
        logger.info(f"Fetching up to {limit} trending products...")

        # Method 1: Official API (Recommended)
        if self._has_api_credentials():
            return self._fetch_via_api(limit)

        # Method 2: Web Scraping (Use with caution)
        else:
            logger.warning("No API credentials found. Using demo data.")
            return self._generate_demo_products(limit)

    def _has_api_credentials(self) -> bool:
        """Check if API credentials are configured"""
        api_config = self.credentials.get('tiktok_shop_api', {})
        return bool(
            api_config.get('app_key') and
            api_config.get('app_secret') and
            api_config.get('access_token')
        )

    def _fetch_via_api(self, limit: int) -> List[Dict]:
        """
        Fetch products using official TikTok Shop API
        Documentation: https://partner.tiktokshop.com/doc/page/262
        """
        api_config = self.credentials.get('tiktok_shop_api', {})

        # API endpoint (example - check official docs)
        base_url = "https://open-api.tiktokglobalshop.com"
        endpoint = f"{base_url}/product/202309/products/search"

        params = {
            'app_key': api_config['app_key'],
            'access_token': api_config['access_token'],
            'page_size': limit,
            'sort_by': 'sales',  # or 'price', 'rating'
            'timestamp': int(time.time())
        }

        # Add signature (required by TikTok Shop API)
        # params['sign'] = self._generate_signature(params, api_config['app_secret'])

        try:
            response = self.session.get(endpoint, params=params, timeout=30)
            response.raise_for_status()
            data = response.json()

            if data.get('code') == 0:
                products = data.get('data', {}).get('products', [])
                return self._parse_api_products(products)
            else:
                logger.error(f"API Error: {data.get('message')}")
                return []

        except Exception as e:
            logger.error(f"Error fetching products via API: {e}")
            return []

    def _parse_api_products(self, raw_products: List[Dict]) -> List[Dict]:
        """Parse and filter products from API response"""
        products = []

        for item in raw_products:
            # Calculate commission
            price = float(item.get('price', 0))
            commission_rate = float(item.get('commission_rate', 0))
            commission_amount = price * (commission_rate / 100)

            # Apply filters
            if not self._meets_criteria(price, commission_rate, item.get('rating', 0)):
                continue

            product = {
                'product_id': item.get('product_id'),
                'name': item.get('product_name'),
                'description': item.get('description', ''),
                'price': price,
                'commission_rate': commission_rate,
                'commission_amount': commission_amount,
                'category': item.get('category_name', 'General'),
                'rating': item.get('rating', 0),
                'image_url': item.get('images', [{}])[0].get('url', ''),
                'affiliate_link': self._generate_affiliate_link(item.get('product_id')),
                'product_url': item.get('product_url', ''),
            }
            products.append(product)

        logger.info(f"Parsed {len(products)} products that meet criteria")
        return products

    def _generate_demo_products(self, limit: int) -> List[Dict]:
        """
        Generate demo products for testing
        Replace this with real API integration
        """
        categories = self.filters.get('categories', ['Electronics', 'Beauty', 'Fashion'])
        demo_products = []

        product_names = [
            "Wireless Bluetooth Earbuds Pro",
            "LED Makeup Mirror with Lights",
            "Portable Phone Charger 20000mAh",
            "Smart Watch Fitness Tracker",
            "Hair Straightener Brush",
            "Water Bottle with Time Marker",
            "Phone Ring Light for TikTok",
            "Resistance Bands Set",
            "Face Roller Jade Stone",
            "Laptop Stand Adjustable",
        ]

        for i in range(min(limit, len(product_names))):
            price = random.uniform(15, 150)
            commission_rate = random.uniform(8, 25)

            product = {
                'product_id': f'DEMO_{i+1:04d}',
                'name': product_names[i],
                'description': f'High-quality {product_names[i].lower()} with excellent reviews!',
                'price': round(price, 2),
                'commission_rate': round(commission_rate, 2),
                'commission_amount': round(price * commission_rate / 100, 2),
                'category': random.choice(categories),
                'rating': round(random.uniform(4.0, 5.0), 1),
                'image_url': f'https://via.placeholder.com/400x400?text={product_names[i][:20]}',
                'affiliate_link': f'https://tiktok.com/shop/product/{i+1}?affiliate=demo',
                'product_url': f'https://tiktok.com/shop/product/{i+1}',
            }
            demo_products.append(product)

        logger.info(f"Generated {len(demo_products)} demo products")
        return demo_products

    def _meets_criteria(self, price: float, commission_rate: float, rating: float) -> bool:
        """Check if product meets filtering criteria"""
        filters = self.filters

        if price < filters.get('min_price', 0):
            return False
        if price > filters.get('max_price', 1000):
            return False
        if commission_rate < filters.get('min_commission_rate', 0):
            return False
        if rating < filters.get('min_rating', 0):
            return False

        return True

    def _generate_affiliate_link(self, product_id: str) -> str:
        """Generate affiliate link for a product"""
        # This would use your TikTok affiliate ID
        # Example format (check TikTok's actual format)
        affiliate_id = self.credentials.get('tiktok', {}).get('affiliate_id', 'YOUR_ID')
        return f"https://www.tiktok.com/shop/product/{product_id}?affiliate={affiliate_id}"

    def download_product_image(self, image_url: str, save_path: Path) -> bool:
        """Download product image"""
        try:
            response = self.session.get(image_url, timeout=30)
            response.raise_for_status()

            with open(save_path, 'wb') as f:
                f.write(response.content)

            logger.info(f"Downloaded image to {save_path}")
            return True

        except Exception as e:
            logger.error(f"Error downloading image: {e}")
            return False

    def search_products(self, query: str, limit: int = 10) -> List[Dict]:
        """Search for specific products by keyword"""
        logger.info(f"Searching for: {query}")
        # Implement search functionality similar to fetch_trending_products
        # but with search query parameter
        return []


# Example Usage
if __name__ == "__main__":
    logging.basicConfig(level=logging.INFO)

    # Demo credentials
    credentials = {
        'tiktok_shop_api': {
            'app_key': '',
            'app_secret': '',
            'access_token': ''
        }
    }

    filters = {
        'min_commission_rate': 10,
        'min_price': 10,
        'max_price': 200,
        'min_rating': 4.0,
        'categories': ['Electronics', 'Beauty', 'Fashion']
    }

    fetcher = ProductFetcher(credentials, filters)
    products = fetcher.fetch_trending_products(limit=5)

    for product in products:
        print(f"\n{product['name']}")
        print(f"  Price: ${product['price']}")
        print(f"  Commission: {product['commission_rate']}% (${product['commission_amount']})")
        print(f"  Rating: {product['rating']}")
